Move files to correct location
