+++ sailinggame/src/main/java/com/group11/model/utility/AStar.java
@@ -1,4 +1,4 @@
-package com.group11.model;
+package com.group11.model.utility;
 
 import java.util.*;
